consolidated logic to copy files to an HTTP output stream
