+++ app/core/src/main/scala/com/lollypop/runtime/datatypes/CLOB.scala
@@ -2,7 +2,7 @@ package com.lollypop.runtime.datatypes
 
 import com.lollypop.runtime.DatabaseObjectNS
 import com.lollypop.runtime.devices.RowCollectionZoo.createTempNS
-import com.lollypop.util.CodecHelper.{RichInputStream, RichReader}
+import com.lollypop.util.IOTools.{RichInputStream, RichReader}
 import com.lollypop.util.ResourceHelper._
 import org.apache.commons.io.IOUtils
