+++ app/core/src/test/scala/com/lollypop/runtime/datatypes/CLOBTest.scala
@@ -2,8 +2,8 @@ package com.lollypop.runtime.datatypes
 
 import com.lollypop.runtime.RuntimeFiles.RecursiveFileList
 import com.lollypop.runtime.instructions.VerificationTools
-import com.lollypop.util.CodecHelper.{RichInputStream, RichReader}
 import com.lollypop.util.ResourceHelper._
+import com.lollypop.util.IOTools.{RichInputStream, RichReader}
 import org.scalatest.funspec.AnyFunSpec
 
 import java.io.{ByteArrayInputStream, File, FileReader, InputStreamReader}
