+++ app/jdbc-driver/src/main/scala/com/lollypop/database/jdbc/JDBCCallableStatement.scala
@@ -3,7 +3,7 @@ package com.lollypop.database.jdbc
 import com.lollypop.database.jdbc.types.JDBCValueConversion._
 import com.lollypop.database.jdbc.types._
 import com.lollypop.runtime.datatypes._
-import com.lollypop.util.CodecHelper.{RichInputStream, RichReader}
+import com.lollypop.util.IOTools.{RichInputStream, RichReader}
 
 import java.io.{InputStream, Reader}
 import java.net.{URI, URL}
